Create seperate test file instead
