+++ docker-app/qfieldcloud/core/tests/test_qfield_file.py
@@ -9,9 +9,7 @@ import requests
 from django.http.response import HttpResponse, HttpResponseRedirect
 from qfieldcloud.authentication.models import AuthToken
 from qfieldcloud.core.geodb_utils import delete_db_and_role
-from qfieldcloud.core.models import ApplyJob, Geodb, Job, PackageJob, Person, Project
-from qfieldcloud.subscription.exceptions import SubscriptionException
-from qfieldcloud.subscription.models import SubscriptionStatus
+from qfieldcloud.core.models import Geodb, Job, PackageJob, Person, Project
 from rest_framework import status
 from rest_framework.test import APITransactionTestCase
 
@@ -29,11 +27,7 @@ class QfcTestCase(APITransactionTestCase):
 
         self.user2 = Person.objects.create_user(username="user2", password="abc123")
 
-        self.token1 = AuthToken.objects.get_or_create(
-            user=self.user1,
-            client_type=AuthToken.ClientType.QFIELD,
-            user_agent="qfield|dev",
-        )[0]
+        self.token1 = AuthToken.objects.get_or_create(user=self.user1)[0]
 
         # Create a project
         self.project1 = Project.objects.create(
@@ -89,79 +83,6 @@ class QfcTestCase(APITransactionTestCase):
         except Exception:
             self.assertTrue(status.is_success(response.status_code), response.content)
 
-    def test_push_file_to_qfield_always_allowed(self):
-        self.client.credentials(HTTP_AUTHORIZATION="Token " + self.token1.key)
-
-        # FIXME TODO 'worker'
-        self.assertEqual(self.token1.client_type, AuthToken.ClientType.QFIELD)
-
-        # Check 'project owner' and 'client' user are same to ensure all scenarios
-        self.assertEqual(self.project1.owner.id, self.token1.user.id)
-        account = self.user1.useraccount
-        subscription = account.current_subscription
-        subscription.status = SubscriptionStatus.INACTIVE_DRAFT
-        subscription.save()
-
-        self.assertFalse(account.current_subscription.is_active)
-
-        plan = subscription.plan
-        plan.storage_mb = 0
-        plan.save()
-        self.assertEqual(account.storage_free_bytes, 0)
-
-        # Create a project that uses all the storage
-        # more_bytes_than_plan = (plan.storage_mb * 1000 * 1000) + 1
-        # Project.objects.create(
-        #     name="p1",
-        #     owner=self.user1,
-        #     file_storage_bytes=more_bytes_than_plan,
-        # )
-
-        # A cross check that no delta apply or package jobs can be created on the project
-        with self.assertRaises(SubscriptionException):
-            PackageJob.objects.create(
-                type=Job.Type.PACKAGE, project=self.project1, created_by=self.user1
-            )
-        with self.assertRaises(SubscriptionException):
-            ApplyJob.objects.create(
-                type=Job.Type.DELTA_APPLY, project=self.project1, created_by=self.user1
-            )
-
-        cur = self.conn.cursor()
-
-        cur.execute(
-            """
-            CREATE TABLE point (
-                id          integer,
-                geometry   geometry(point, 2056)
-            );
-            """
-        )
-
-        self.conn.commit()
-
-        cur.execute(
-            """
-            INSERT INTO point(id, geometry)
-            VALUES(1, ST_GeomFromText('POINT(2725505 1121435)', 2056));
-            """
-        )
-        self.conn.commit()
-
-        # Add the qgis project
-        file = testdata_path("delta/project2.qgs")
-        response = self.client.post(
-            "/api/v1/files/{}/project.qgs/".format(self.project1.id),
-            {"file": open(file, "rb")},
-            format="multipart",
-        )
-        self.assertTrue(status.is_success(response.status_code))
-
-        response = self.client.post(
-            "/api/v1/qfield-files/export/{}/".format(self.project1.id)
-        )
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
     def test_list_files_for_qfield(self):
         self.client.credentials(HTTP_AUTHORIZATION="Token " + self.token1.key)
 
@@ -198,7 +119,6 @@ class QfcTestCase(APITransactionTestCase):
         response = self.client.post(
             "/api/v1/qfield-files/export/{}/".format(self.project1.id)
         )
-
         self.assertTrue(status.is_success(response.status_code))
 
         # Wait for the worker to finish
